third_party/blink/web_tests/external/wpt/wasm/jsapi/table/grow-reftypes.tentative.any-expected.txt

This is a testharness.js-based test.
FAIL Grow with exported-function argument assert_equals: expected (function) function "function 0() { [native code] }" but got (object) null
FAIL Grow with non-function argument assert_throws_js: function "() => table.grow(2, {})" did not throw
FAIL Grow with JS-function argument assert_throws_js: function "() => table.grow(2, () => true)" did not throw
Harness: the test ran to completion.
